requirements.txt
psycopg2-binary==2.9.9
openai==1.12.0
tabulate==0.9.0 